Add files via upload
This file displays the tests to show that the doubly linked list is properly being used in terms of memory, and the functionality of the project is properly being executed. This can be run by its own. UseThis command to compile : g++ -o work work.cpp
and this to run : ./work